+++ CHANGELOG.rst
@@ -7,13 +7,22 @@ smaht-portal
 Change Log
 ----------
 
-1.4.9
+1.5.1
 =====
+
 `PR 538: feat: updates to documentation pages <https://github.com/smaht-dac/smaht-portal/pull/538>`_
 
 * Updates image on the account creation page
 * Fix typos/spacing
 
+
+1.5.0
+=====
+
+`PR 545: Refactor Cypress tests for role-based testing <https://github.com/smaht-dac/smaht-portal/pull/545>`
+
+* Refactor Cypress tests to add role-based testing capabilities
+* Add new Cypress tests for Documentation page, Data Overview page, and File Overview page
 
 1.4.8
 =====